wip: refactoring results for cakes and pancakes
