+++ crates/results/cakes/old_src/main.rs
@@ -1,0 +1,429 @@
+#![deny(clippy::correctness)]
+#![warn(
+    missing_docs,
+    clippy::all,
+    clippy::suspicious,
+    clippy::style,
+    clippy::complexity,
+    clippy::perf,
+    clippy::pedantic,
+    clippy::nursery,
+    clippy::unwrap_used,
+    clippy::expect_used,
+    clippy::panic,
+    clippy::cast_lossless
+)]
+#![doc = include_str!("../README.md")]
+
+use std::path::PathBuf;
+
+use abd_clam::{
+    adapter::{ParAdapter, ParBallAdapter},
+    cakes::{Algorithm, CodecData, Decompressible, OffBall, SquishyBall},
+    partition::ParPartition,
+    BalancedBall, Ball, Cluster, Dataset, FlatVec, MetricSpace,
+};
+use clap::Parser;
+
+mod genomic;
+mod member_set;
+mod metrics;
+mod sequence;
+mod sets;
+mod tables;
+
+use distances::Number;
+use member_set::MemberSet;
+use metrics::SetDistance;
+
+/// The Vector of held-out queries
+pub type QueriesGen = Vec<(String, String)>;
+/// The type of the compressible dataset.
+pub type CoGen = FlatVec<String, u32, String>;
+/// The type of the ball tree over the compressible dataset.
+#[allow(dead_code)]
+type BGen = Ball<String, u32, CoGen>;
+/// The type of the compressed, decompressible dataset.
+#[allow(dead_code)]
+type DecGen = CodecData<String, u32, String>;
+/// The type of the squishy ball tree.
+#[allow(dead_code)]
+type SBGen = SquishyBall<String, u32, CoGen, DecGen, BGen>;
+
+/// The Vector of held-out queries
+pub type QueriesSet = Vec<(usize, MemberSet)>;
+/// The type of the compressible dataset.
+pub type CoSet = FlatVec<MemberSet, f32, usize>;
+/// The type of the ball tree over the compressible dataset.
+type BSet = Ball<MemberSet, f32, CoSet>;
+/// The type of the compressed, decompressible dataset.
+type DecSet = CodecData<MemberSet, f32, usize>;
+/// The type of the squishy ball tree.
+type SBSet = SquishyBall<MemberSet, f32, CoSet, DecSet, BSet>;
+
+/// Simple program to greet a person
+#[derive(Parser, Debug)]
+#[command(version, about, long_about = None)]
+struct Args {
+    /// Dataset type
+    #[arg(short, long)]
+    // dataset: genomic::GenomicDataset,
+    dataset: sets::SetDataset,
+
+    /// The number of queries to hold out.
+    #[arg(short, long)]
+    num_queries: usize,
+
+    /// Path to the input directory.
+    #[arg(short, long)]
+    inp_dir: PathBuf,
+
+    /// Path to the output directory.
+    #[arg(short, long)]
+    out_dir: Option<PathBuf>,
+}
+
+#[allow(clippy::too_many_lines)]
+fn main() -> Result<(), String> {
+    let args = Args::parse();
+
+    mt_logger::mt_new!(
+        Some("cakes-results.log"),
+        mt_logger::Level::Debug,
+        mt_logger::OutputStream::Both
+    );
+
+    let out_dir = args.out_dir.unwrap_or_else(|| {
+        let out_dir = args.inp_dir.join("cakes-results");
+        if !out_dir.exists() {
+            #[allow(clippy::expect_used)]
+            std::fs::create_dir(&out_dir).expect("Failed to create output directory");
+        }
+        out_dir
+    });
+
+    // Check that the directories exist
+    for dir in [&args.inp_dir, &out_dir] {
+        if !dir.exists() {
+            return Err(format!("{dir:?} does not exist!"));
+        }
+        if !dir.is_dir() {
+            return Err(format!("{dir:?} is not a directory!"));
+        }
+    }
+
+    let inp_dir = args.inp_dir.canonicalize().map_err(|e| e.to_string())?;
+    mt_logger::mt_log!(mt_logger::Level::Info, "Input directory: {inp_dir:?}");
+
+    let out_dir = out_dir.canonicalize().map_err(|e| e.to_string())?;
+    mt_logger::mt_log!(mt_logger::Level::Info, "Output directory: {out_dir:?}");
+
+    let ball_path = out_dir.join(args.dataset.ball_file());
+    let flat_vec_path = out_dir.join(args.dataset.flat_file());
+    let queries_path = out_dir.join(args.dataset.queries_file());
+
+    let squishy_ball_path = out_dir.join(args.dataset.squishy_ball_file());
+    let codec_data_path = out_dir.join(args.dataset.compressed_file());
+
+    let extension = "csv";
+    let ball_table_path = out_dir.join(args.dataset.ball_table("ball", extension));
+    let pre_trim_table_path = out_dir.join(args.dataset.ball_table("pre_trim", extension));
+    let squishy_ball_table_path = out_dir.join(args.dataset.ball_table("squishy_ball", extension));
+
+    // Read the dataset
+    let (data, queries) = if flat_vec_path.exists() {
+        let data: CoSet = bincode::deserialize_from(std::fs::File::open(&flat_vec_path).map_err(|e| e.to_string())?)
+            .map_err(|e| e.to_string())?;
+
+        let queries: QueriesSet =
+            bincode::deserialize_from(std::fs::File::open(&queries_path).map_err(|e| e.to_string())?)
+                .map_err(|e| e.to_string())?;
+
+        (data, queries)
+    } else {
+        // let (data, queries): (CoGen, QueriesGen) = args.dataset.read_fasta(&inp_dir, args.num_queries)?;
+        let (data, queries): (CoSet, QueriesSet) = args.dataset.read_raw(&inp_dir)?;
+
+        bincode::serialize_into(std::fs::File::create(&flat_vec_path).map_err(|e| e.to_string())?, &data)
+            .map_err(|e| e.to_string())?;
+
+        bincode::serialize_into(
+            std::fs::File::create(&queries_path).map_err(|e| e.to_string())?,
+            &queries,
+        )
+        .map_err(|e| e.to_string())?;
+
+        (data, queries)
+    };
+
+    let data = {
+        let mut data = data;
+        data.set_metric(SetDistance::Jaccard.metric());
+        data
+    };
+
+    mt_logger::mt_log!(
+        mt_logger::Level::Info,
+        "Working with {:?} Dataset with {} sequences in {:?} dims.",
+        args.dataset,
+        data.cardinality(),
+        data.dimensionality_hint()
+    );
+
+    mt_logger::mt_log!(mt_logger::Level::Info, "Holding out {} queries", queries.len());
+
+    let ball: BSet = if ball_path.exists() {
+        bincode::deserialize_from(std::fs::File::open(&ball_path).map_err(|e| e.to_string())?)
+            .map_err(|e| e.to_string())?
+    } else {
+        let mut max_depth = 0;
+        let depth_delta = abd_clam::MAX_RECURSION_DEPTH;
+        let seed = Some(42);
+
+        let indices = (0..data.cardinality()).collect::<Vec<_>>();
+        let mut ball = BalancedBall::par_new(&data, &indices, 0, seed);
+
+        loop {
+            max_depth += depth_delta;
+            let criteria =
+                |c: &BalancedBall<_, _, _>| (c.depth() < max_depth && c.radius() > 0.75) || c.cardinality() > 1_000;
+            ball.par_partition_further(&data, &criteria, seed);
+
+            // If there are no leaves at the current maximum depth, break
+            if !ball.leaves().into_iter().any(|c| c.depth() == max_depth) {
+                break;
+            }
+        }
+
+        let mut ball = BSet::par_from_balanced_ball(ball);
+
+        for leaf in ball.leaves_mut() {
+            max_depth = leaf.depth();
+            loop {
+                max_depth += depth_delta;
+                let criteria = |c: &BSet| c.depth() < max_depth;
+                leaf.par_partition_further(&data, &criteria, seed);
+
+                // If there are no leaves at the current maximum depth, break
+                if !leaf.leaves().into_iter().any(|c| c.depth() == max_depth) {
+                    break;
+                }
+            }
+        }
+
+        bincode::serialize_into(std::fs::File::create(&ball_path).map_err(|e| e.to_string())?, &ball)
+            .map_err(|e| e.to_string())?;
+
+        ball
+    };
+
+    tables::write_ball_csv(&ball, &ball_table_path)?;
+
+    let subtree_cardinality = ball.subtree().len();
+    mt_logger::mt_log!(mt_logger::Level::Info, "BallTree has {subtree_cardinality} clusters.");
+
+    let (squishy_ball, perm_data) = {
+        let (off_ball, data) = OffBall::par_from_ball_tree(ball.clone(), data.clone());
+        let mut pre_trim_ball = SquishyBall::par_adapt_tree_iterative(off_ball, None);
+        pre_trim_ball.par_set_costs(&data);
+        tables::write_squishy_ball_csv(&pre_trim_ball, &pre_trim_table_path)?;
+
+        let pre_trim_subtree_cardinality = pre_trim_ball.subtree().len();
+        mt_logger::mt_log!(
+            mt_logger::Level::Info,
+            "Pre-trimmed SquishyBall has {pre_trim_subtree_cardinality} clusters."
+        );
+
+        (pre_trim_ball, data)
+    };
+
+    let metadata = data.metadata().to_vec();
+    let (squishy_ball, codec_data): (SBSet, DecSet) = if squishy_ball_path.exists() && codec_data_path.exists() {
+        core::mem::drop(squishy_ball);
+
+        let squishy_ball: SBSet =
+            bincode::deserialize_from(std::fs::File::open(&squishy_ball_path).map_err(|e| e.to_string())?)
+                .map_err(|e| e.to_string())?;
+
+        let squishy_ball_subtree_cardinality = squishy_ball.subtree().len();
+        mt_logger::mt_log!(
+            mt_logger::Level::Info,
+            "SquishyBall has {squishy_ball_subtree_cardinality} clusters."
+        );
+
+        let mut codec_data: DecSet =
+            bincode::deserialize_from(std::fs::File::open(&codec_data_path).map_err(|e| e.to_string())?)
+                .map_err(|e| e.to_string())?;
+
+        let num_leaf_bytes = codec_data.leaf_bytes().len();
+        mt_logger::mt_log!(mt_logger::Level::Info, "CodecData has {num_leaf_bytes} leaf bytes.");
+
+        codec_data.set_metric(data.metric().clone());
+
+        (squishy_ball, codec_data)
+    } else if squishy_ball_path.exists() {
+        core::mem::drop(squishy_ball);
+
+        let squishy_ball: SBSet =
+            bincode::deserialize_from(std::fs::File::open(&squishy_ball_path).map_err(|e| e.to_string())?)
+                .map_err(|e| e.to_string())?;
+
+        let squishy_ball_subtree_cardinality = squishy_ball.subtree().len();
+        mt_logger::mt_log!(
+            mt_logger::Level::Info,
+            "SquishyBall has {squishy_ball_subtree_cardinality} clusters."
+        );
+
+        // let squishy_ball = squishy_ball.with_metadata_type::<usize>();
+        let codec_data = CodecData::par_from_compressible(&data, &squishy_ball);
+
+        let num_leaf_bytes = codec_data.leaf_bytes().len();
+        mt_logger::mt_log!(mt_logger::Level::Info, "CodecData has {num_leaf_bytes} leaf bytes.");
+
+        // let squishy_ball = squishy_ball.with_metadata_type::<String>();
+        let codec_data: DecSet = codec_data.with_metadata(metadata)?;
+
+        bincode::serialize_into(
+            std::fs::File::create(&codec_data_path).map_err(|e| e.to_string())?,
+            &codec_data,
+        )
+        .map_err(|e| e.to_string())?;
+
+        (squishy_ball, codec_data)
+    } else {
+        let mut squishy_ball = squishy_ball;
+        squishy_ball.trim();
+
+        // let squishy_ball = squishy_ball.with_metadata_type::<usize>();
+        let codec_data = CodecData::par_from_compressible(&perm_data, &squishy_ball);
+
+        core::mem::drop(perm_data);
+
+        // let squishy_ball: SBGen = squishy_ball.with_metadata_type::<String>();
+        let codec_data: DecSet = codec_data.with_metadata(metadata)?;
+
+        let squishy_ball_subtree_cardinality = squishy_ball.subtree().len();
+        mt_logger::mt_log!(
+            mt_logger::Level::Info,
+            "SquishyBall has {squishy_ball_subtree_cardinality} clusters."
+        );
+
+        let num_leaf_bytes = codec_data.leaf_bytes().len();
+        mt_logger::mt_log!(mt_logger::Level::Info, "CodecData has {num_leaf_bytes} leaf bytes.");
+
+        bincode::serialize_into(
+            std::fs::File::create(&squishy_ball_path).map_err(|e| e.to_string())?,
+            &squishy_ball,
+        )
+        .map_err(|e| e.to_string())?;
+
+        bincode::serialize_into(
+            std::fs::File::create(&codec_data_path).map_err(|e| e.to_string())?,
+            &codec_data,
+        )
+        .map_err(|e| e.to_string())?;
+
+        (squishy_ball, codec_data)
+    };
+
+    tables::write_squishy_ball_csv(&squishy_ball, &squishy_ball_table_path)?;
+
+    mt_logger::mt_flush!().map_err(|e| e.to_string())?;
+
+    // Note: Starting search benchmarks here
+
+    let (_, queries): (Vec<_>, Vec<_>) = queries.into_iter().unzip();
+    // let (data, codec_data) = {
+    //     let metric = StringDistance::Levenshtein.metric();
+    //     let mut data = data;
+    //     data.set_metric(metric.clone());
+    //     let mut codec_data = codec_data;
+    //     codec_data.set_metric(metric);
+    //     (data, codec_data)
+    // };
+
+    let algorithms: Vec<Algorithm<f32>> = {
+        let mut algorithms = Vec::new();
+
+        for radius in [1, 5, 10] {
+            let radius = radius.as_f32() / 100.0;
+            algorithms.push(Algorithm::RnnLinear(radius));
+            algorithms.push(Algorithm::RnnClustered(radius));
+        }
+
+        for k in [1, 10, 100] {
+            algorithms.push(Algorithm::KnnLinear(k));
+            // algorithms.push(Algorithm::KnnRepeatedRnn(k, 2.0));
+            algorithms.push(Algorithm::KnnBreadthFirst(k));
+            algorithms.push(Algorithm::KnnDepthFirst(k));
+        }
+
+        algorithms
+    };
+
+    mt_logger::mt_log!(
+        mt_logger::Level::Info,
+        "Starting search benchmarks on {} algorithms ...",
+        algorithms.len()
+    );
+
+    // TODO: Remove this limit
+    // let queries = &queries[..100];
+    let queries = &queries;
+
+    for (i, alg) in algorithms.iter().enumerate() {
+        mt_logger::mt_log!(
+            mt_logger::Level::Info,
+            "\n\nStarting {} Search ({}/{}) on Ball and FlatVec ...",
+            alg.name(),
+            i + 1,
+            algorithms.len()
+        );
+        let start = std::time::Instant::now();
+        let hits = alg.par_batch_par_search(&data, &ball, queries);
+        let ball_end = start.elapsed().as_secs_f32();
+        mt_logger::mt_log!(
+            mt_logger::Level::Info,
+            "Finished {} Search ({}/{}) on Ball and FlatVec in {ball_end:.6} seconds.",
+            alg.name(),
+            i + 1,
+            algorithms.len()
+        );
+        let mean_num_hits = abd_clam::utils::mean::<_, f32>(&hits.into_iter().map(|h| h.len()).collect::<Vec<_>>());
+        mt_logger::mt_log!(mt_logger::Level::Info, "Average number of hits was {mean_num_hits:.6}.");
+
+        mt_logger::mt_log!(
+            mt_logger::Level::Info,
+            "Starting {} Search ({}/{}) on SquishyBall and CodecData ...",
+            alg.name(),
+            i + 1,
+            algorithms.len()
+        );
+        let start = std::time::Instant::now();
+        let hits = alg.par_batch_par_search(&codec_data, &squishy_ball, queries);
+        let comp_end = start.elapsed().as_secs_f32();
+        mt_logger::mt_log!(
+            mt_logger::Level::Info,
+            "Finished {} Search ({}/{}) on SquishyBall and CodecData in {comp_end:.6} seconds.",
+            alg.name(),
+            i + 1,
+            algorithms.len()
+        );
+        let mean_num_hits = abd_clam::utils::mean::<_, f32>(&hits.into_iter().map(|h| h.len()).collect::<Vec<_>>());
+        mt_logger::mt_log!(mt_logger::Level::Info, "Average number of hits was {mean_num_hits:.6}.");
+
+        let ratio = comp_end / ball_end;
+        mt_logger::mt_log!(
+            mt_logger::Level::Info,
+            "Speed ratio of Ball over SquishyBall was {ratio:.2}x."
+        );
+    }
+
+    mt_logger::mt_log!(
+        mt_logger::Level::Info,
+        "Finished search benchmarks on {} algorithms ...",
+        algorithms.len()
+    );
+
+    mt_logger::mt_flush!().map_err(|e| e.to_string())
+}
